Writes data to a Excel spreadsheet
For Raspberry Pi 4. Saves data from the BMP280 senor to a spreadsheet in Excel format. I use LibraOffice and the code saves it to the document weater.xlsx in Raspberry Pi 4 Documents directory.
Please note: I am in no way a programmer so the code is experimental and is put together from tutorials from Internet